refactor(agent): 迁移条件与表达式求值到 workflow/condition 子包
