+++ internal/agent/workflow/condition/expression.go
@@ -1,4 +1,4 @@
-package agent
+package condition
 
 import (
 	"encoding/json"
